refactor: organize modules
