+++ src/preview/PreviewProvider.tsx
@@ -1,7 +1,7 @@
 import {GroqStoreProvider, type GroqStoreProviderProps} from '@sanity/preview-kit/groq-store'
 import type {ReactNode} from 'react'
 
-export type PreviewProviderProps = GroqStoreProviderProps & {children: ReactNode}
+type PreviewProviderProps = GroqStoreProviderProps & {children: ReactNode}
 
 export function PreviewProvider(props: PreviewProviderProps) {
   const {children, ...rest} = props
